refacto: 범용적으로 사용할 수 있게 BrakeTextFieldView 수정 및 BrakeValidatedTextFieldView 생성
